refactor: refactoring bingads poll
